+++ prince-archiver/prince_archiver/service_layer/streams.py
@@ -9,7 +9,7 @@ from prince_archiver.adapters.streams import (
 )
 
 from .exceptions import InvalidStreamMessage
-from .messages import ExportedImagingEvent, ImagingEventStream
+from .messages import ExportedImagingEvent, NewImagingEvent
 
 LOGGER = logging.getLogger(__name__)
 
@@ -25,17 +25,17 @@ class Group(StrEnum):
 
 
 class Message(AbstractOutgoingMessage):
-    def __init__(self, data: ImagingEventStream):
+    def __init__(self, data: NewImagingEvent):
         self.data = data
 
     def fields(self) -> dict:
         return self.data.model_dump(mode="json", round_trip=True)
 
 
-class IncomingMessage(AbstractIncomingMessage[ImagingEventStream]):
-    def processed_data(self) -> ImagingEventStream:
+class IncomingMessage(AbstractIncomingMessage[NewImagingEvent]):
+    def processed_data(self) -> NewImagingEvent:
         try:
-            return ImagingEventStream(
+            return NewImagingEvent(
                 **{k.decode(): v.decode() for k, v in self.raw_data.items()},
             )
         except ValidationError as exc:
