+++ prince-archiver/tests/unit/entrypoints/test_state_manager.py
@@ -1,47 +1,47 @@
-from datetime import date
 from unittest.mock import AsyncMock
-from uuid import UUID
 
 import pytest
 from aio_pika.abc import AbstractIncomingMessage
 
 from prince_archiver.entrypoints.state_manager.consumers import SubscriberMessageHandler
-from prince_archiver.service_layer.external_dto import Archive, UpdateArchiveEntries
 from prince_archiver.service_layer.messagebus import MessageBus
-from prince_archiver.service_layer.messages import AddDataArchiveEntry, ArchiveMember
+from prince_archiver.service_layer.messages import (
+    AddDataArchiveEntry,
+    NewDataArchiveEntries,
+)
 
 
-@pytest.fixture(name="update_archive_entries_msg")
-def fixture_update_archive_entries_msg() -> UpdateArchiveEntries:
-    return UpdateArchiveEntries(
-        job_id=UUID("a136586a44e6417eb707e10d9795b1f9"),
-        date=date(2000, 1, 1),
+@pytest.fixture(name="new_data_archive_entries")
+def fixture_new_data_archive_entries_msg() -> NewDataArchiveEntries:
+    return NewDataArchiveEntries(
+        job_id="a136586a44e6417eb707e10d9795b1f9",
+        date="2000-01-01",
         archives=[
-            Archive(
-                id=UUID("dabfa4a3051c4e47a736e9d12e38b05a"),
-                path="test_path",
-                src_keys=["test/a"],
-            )
+            {
+                "id": "dabfa4a3051c4e47a736e9d12e38b05a",
+                "path": "test_path",
+                "src_keys": ["test/a"],
+            }
         ],
     )
 
 
 async def test_subscriber_message_handler(
-    update_archive_entries_msg: UpdateArchiveEntries,
+    new_data_archive_entries: NewDataArchiveEntries,
 ):
     messagebus = AsyncMock(MessageBus)
     handler = SubscriberMessageHandler(messagebus_factory=lambda: messagebus)
 
     incoming_message = AsyncMock(
         AbstractIncomingMessage,
-        body=update_archive_entries_msg.model_dump_json().encode(),
+        body=new_data_archive_entries.model_dump_json().encode(),
     )
 
     expected_msg = AddDataArchiveEntry(
-        id=UUID("dabfa4a3051c4e47a736e9d12e38b05a"),
-        job_id=UUID("a136586a44e6417eb707e10d9795b1f9"),
+        id="dabfa4a3051c4e47a736e9d12e38b05a",
+        job_id="a136586a44e6417eb707e10d9795b1f9",
         path="test_path",
-        members=[ArchiveMember(src_key="test/a", member_key="a")],
+        members=[{"src_key": "test/a", "member_key": "a"}],
     )
 
     await handler(incoming_message)
