+++ prince-archiver/tests/unit/handlers/test_persist_export.py
@@ -1,16 +1,11 @@
-from datetime import UTC, datetime
 from uuid import uuid4
 
 import pytest
 
-from prince_archiver.definitions import Algorithm
 from prince_archiver.domain.models import ImagingEvent
 from prince_archiver.service_layer.exceptions import ServiceLayerException
 from prince_archiver.service_layer.handlers.state import persist_imaging_event_export
-from prince_archiver.service_layer.messages import (
-    Checksum,
-    ExportedImagingEvent,
-)
+from prince_archiver.service_layer.messages import ExportedImagingEvent
 
 from .utils import MockUnitOfWork
 
@@ -21,10 +16,10 @@ async def test_persist_imaging_event_successful(
 ):
     msg = ExportedImagingEvent(
         ref_id=unexported_imaging_event.ref_id,
-        checksum=Checksum(hex="test", algorithm=Algorithm.SHA256),
+        checksum={"hex":"test", "algorithm": "sha256"},
         size=1024,
         key="target",
-        timestamp=datetime(2000, 1, 1, tzinfo=UTC),
+        timestamp="2000-01-01T00:00:00+00:00",
     )
 
     await persist_imaging_event_export(msg, uow)
@@ -38,10 +33,10 @@ async def test_persist_imaging_event_successful(
 async def test_persist_imaging_event_non_existent_reference():
     msg = ExportedImagingEvent(
         ref_id=uuid4(),
-        checksum=Checksum(hex="test", algorithm=Algorithm.SHA256),
+        checksum={"hex":"test", "algorithm": "sha256"},
         size=1024,
         key="target",
-        timestamp=datetime(2000, 1, 1, tzinfo=UTC),
+        timestamp="2000-01-01T00:00:00+00:00",
     )
 
     with pytest.raises(ServiceLayerException):
